requirements.txt
Pillow
pylzham
zstandard
pyliblzfse
texture2ddecoder
